+++ packages/mambajs-core/src/helper.ts
@@ -1,6 +1,6 @@
 import { FilesData, IUnpackJSAPI } from '@emscripten-forge/untarjs';
 import { parse } from 'yaml';
-import { ISolvedPackages, TSharedLibs } from './types';
+import { DEFAULT_CHANNEL_PRIORITY, DEFAULT_CHANNELS, ILock, TSharedLibs } from './types';
 
 export function parseEnvYml(envYml: string) {
   const data = parse(envYml);
@@ -450,18 +450,79 @@ export function getCondaMetaFile(
   return {};
 }
 
-export function splitPipPackages(installed?: ISolvedPackages) {
-  const installedCondaPackages: ISolvedPackages = {};
-  const installedPipPackages: ISolvedPackages = {};
-  if (installed) {
-    Object.keys(installed).filter((filename: string) => {
-      const pkg = installed[filename];
-      if (pkg.repo_name !== 'PyPi') {
-        installedCondaPackages[filename] = pkg;
-      } else {
-        installedPipPackages[filename] = pkg;
-      }
-    });
+export function formatChannels(channels?: string[]): Pick<ILock, 'channels' | 'channel_priority'> {
+  if (!channels || !channels.length) {
+    return {
+      channels: DEFAULT_CHANNELS,
+      channel_priority: DEFAULT_CHANNEL_PRIORITY
+    };
   }
-  return { installedCondaPackages, installedPipPackages };
+
+  const formattedChannels: Pick<ILock, 'channels' | 'channel_priority'> = {
+    channels: {},
+    channel_priority: []
+  };
+
+  // Returns the default channel name if it's a default one, otherwise null
+  const getDefaultChannel = (urlOrName: string): {
+    name: string,
+    channel: ILock['channels'][keyof ILock['channels']]
+  } | null => {
+    // Check if it's a known channel alias
+    if (DEFAULT_CHANNEL_PRIORITY.includes(urlOrName)) {
+      return {
+        name: urlOrName,
+        channel: DEFAULT_CHANNELS[urlOrName]
+      };
+    }
+
+    // If it's a url, check if it matches a default channel mirror
+    Object.keys(DEFAULT_CHANNELS).forEach(name => {
+      const mirrors = DEFAULT_CHANNELS[name];
+      mirrors.forEach(mirror => {
+        if (urlOrName === mirror.url) {
+          return {
+            name,
+            channel: mirrors
+          }
+        }
+      });
+    });
+
+    return null;
+  }
+
+  const pushChannel = (channel: string) => {
+    // Cleanup trailing url slash
+    channel = cleanUrl(channel);
+
+    // If it's defaults, push all default channels
+    if (channel === 'defaults') {
+      DEFAULT_CHANNEL_PRIORITY.forEach(pushChannel);
+      return;
+    }
+
+    // If it's one of the default channels and it's not included yet, add it
+    const asDefaultChannel = getDefaultChannel(channel);
+    if (asDefaultChannel && !formattedChannels.channel_priority.includes(asDefaultChannel.name)) {
+      formattedChannels.channel_priority.push(asDefaultChannel.name);
+      formattedChannels.channels[asDefaultChannel.name] = asDefaultChannel.channel;
+      return;
+    }
+
+    // Otherwise, add it if it's not included yet
+    if (!formattedChannels.channel_priority.includes(channel)) {
+      formattedChannels.channel_priority.push(channel);
+      formattedChannels.channels[channel] = [{url: channel, protocol: 'https'}];
+      return;
+    }
+  }
+
+  channels?.forEach(pushChannel);
+
+  return formattedChannels;
+}
+
+export function cleanUrl(url: string): string {
+  return url.replace(/[\/\s]+$/, '');
 }
